Change default mode to provide an init
The default is to expect that the test binary can be used as init. If
not, the "-wrap" flag and mode can be used.

Since the ad-hoc approach should be as simple as possible and not
require any additional support by the test binary, this weems to be
reverse than it should be.

This commit makes the wrapped init mode the default. If the test binary
has init capabilities itself, the pidonetest needs to be called with
the new "-standalone" flag.