Refactor sysinit entry function
In order to provide the convenient setup and teardown operations but
allow more flexibility, the Run function now takes a function, that is
executed when the system has been prepared. Handling the return code
communication is still done by run, so the given function does not need
to do any system operations.