Refactor ELF shared object resolving
Instead of trying to resolve shared objects ourself, which is quite
fragile, use the actual shared object resolving by the actual interpreter.
So, do the same thing as the "ldd" script does, except that we parse the
actual interpreter from the binary instead of trying a series of known
interpreter.