Fix lib path symbolic link evaluation
The path to the lib added to the initramfs should be the canonical one,
since the interpreter has only this one hard coded. Because of this the
path needs to be resolved. The lib file itself needs to have its name
preserved, so it can be matched correctly.

This commit fixes the symbolic link evaluation so only the directory is
resolved but not the file.