Improve guest output parsing
Only kernel panics are handled specifically and are suppressed if an
return code of the init was found. Out of memory conditions are not
handled well. They led to a return code -1 (255 from virtrun).

This commit changes the stdout parsing by also handling oom messages. It
also introduces specific errors for kernel panics and oom conditions. A
magic return code of 125 for kernel panics is no longer returned.
Instead, the caller can handle the new specific error in that way if
desired.