Pass additional files by flag
It is not possible to pass additional files when using virtrun with
"go test -exec" directly.

This commit adds the repeatable flag "-addFile" to pass additional
files. It also removes the handling for optional additional files as
positional arguments, since it made the interface more complicated to
understand and to handle.

Closes #2