Improve error and return code printing in guest
This commit changes the return code printing, so it is always printed,
even in case of setup errors. Also, it prints the error in that case and
ensures a non zero return code is printed if an error occurred. The
return code must be set to zero explicitly by the caller.