Move pre-compiled inits into seperate package
The pre-compiled init programs are in the initramfs package. They
are built on some assumptions of the main virtrun binary. The
initramfs constructors also had to be built around this assumptions.

This commit moves the pre-compiled inits into their own package. The
initramfs constructor is changed so it can handle regular or embedded
files. The main binary is responsible for providing the proper embedded
file. This way, the initramfs is more generic  and does not care for
architectures or other specifics.