fix: cleanup host test programs
This commit does some cleanup for the integration test programs.
It hardens the panic program by adding a check if it is PID 1 so it can
not shutdown the host system accidentally. This requires being run in
virtrun standalone mode.

Explicit panics for invalid input to the other programs are added.