refactor(initramfs): combine virtual and regular file
Passing an already opened file requires an extra handler, since the
regular file expects a path that can be opened before copying it.

This commit merges the virtual file handling into the regular file by
changing the regular file to be constructed with a function to open the
file when necessary. With this, files from various file systems can be
passed or even already opened files