refactor(initramfs): change into fs.FS
This commit changes the "initramfs" package completely by refactoring it
into a "fs.FS". The "initramfs.FS" is tested with "fstest.TestFS" for
proper implementation of "fs.FS" and "fs.File".

It changes the "Writer" into a "FileWriter" working with any "fs.File".

The "Ldd" implementation and helpers for collecting libraries are moved
to the "sys" package since it is is not related to initramfs directly.

The virtrun's packages use is refactored to reflect those changes.
Readability of the initramfs construction is improved.

With this, the components are much more decoupled, which should allow for
improved testability.