refactor(initramfs): add FS writing to cpio writer
The standard library's archive packages' writers provide an AddFS(fs.FS)
method. It copies from the given fs.FS to the archive. The WriteFS
function in the initramfs package does exactly the same.

In order to match the standard library's archive packages APIs, this
commit integrates the separate WriteFS function as method to the CPIO
writer.