feat(virtrun): set defaults by main binary arch
The architecture for picking the required QEMU defaults is read from
environment variables. Either from a specific one or from cross using
the GOARCH environment variable. Since the binary must be of correct
architecture anyway, this is kinda redundant.

This commit removes the environment variables handling and sets the
defaults based on the architecture of the main binary. This way they can
not be printed as flag defaults anymore, though. Use the `-debug` flag
to see which values are picked.