refactor(qemu): use unix sockets instead of pipes
IO pipes requires passing file descriptors and manual closing. QEMU
chardevs also provide a unix socket backend. With this, the connection
is more independent. Also, since only one connection is made by the QEMU
process, the processor can terminate as soon as the connection closes.
With file descriptors this needed to be done ourself.

This commit replaces the pipe based connection of the additional
consoles by an abstract unix domain socket based setup. Abstract unix
sockets are not created as real files, but exist only in the abstract
name space.