tests(integration): integration tests in packages
The integration tests are all organized in the testing directory. Thus,
they do not live right next to the code they test. In order to have
the tests right next to the code they test, this commit splits the guest
tests. It adds integration tests for the sysinit package, which requires
testing in virtrun in order to test the proper functionality of the
helpers. It also moves the guest tests that actually test the init progs
are right in that package. The CI config is adjusted so the test binary
is used for both of these tests.