feat!(sysinit): provide setup tasks as building blocks
The setup steps are hard coded and just configurable. This makes the
library not very flexible. Also, the exit code communication depends
heavily on both sides knowing how to write/parse the exit code line.

This commit redesigns the package API in order to provide more
flexibility and more robust integration. Instead of a configuration the
Main function takes just simple functions that are run in order and an
error handler that is always run before termination. It just ensures
proper system shutdown. Also, it recovers panics gracefully.

The setup tasks are provided by individual factory functions that take
the users input. The user may pass as many custom setup func and actual
business logic as desired.

In order to handle exit code communication more robust, the package
provides an ExitCodeIdentifier. It specifies the prefix of the exit code
line and provides methods for printing and parsing such a line. It
provides a method that can be used as error handler function with the
main function.

The qemu package is adjusted to take a scan function matching the
signature of the ExitCodeIdentifier.Sscan function. This way the same
tested struct can be used on producer and consumer side.

BREAKING CHANGE: changes complete API of sysinit package