test(sysinit): fix essential file systems
For communicating with the host, the /dev files system is necessary.
Also, /tmp should be present ans usable for the output dir.

This commit adds /dev, /proc, and /tmp as file systems that are mounted
for the integration test. It removes the conflicting temporary mount
helper and adjusts the tests to test with other file systems.