test(cmd): pass essential flags directly
The integration test requires a kernel path so it can actually run. A
dedicated flag for the test is used for that.

Since changing to calling cmd.Run the flags are passed as virtrun flags,
though. This allows to pass through the regular flags.

This commit removes the dedicated test flags and just passes any
additional flags to the tested command directly.