refactor(sysinit): improve exit code scanning
The exit code identifier need to start at the beginning of the line.
Because of this an additional newline is added when printing it just to
make it more resilient against simultaneous printers.

This commit changes the scanning so it does not require the identifier
to be at the beginning of the line.