feat!: transmit data encoded
Output into files are supported by using additional serial consoles.
They are line oriented, though and this creates issues like wrong line
endings, that require removal on the host side. Transmitting binary data
is also tricky, because bytes may be discarded erroneously due to that
post processing.

This commit changes the data transmission for additional files to be
base64 encoded. This allows to transmit any kind of data (binary or
text) safely and correctly.

BREAKING: The guest process is supposed to write into the fifo files
provided by the init instead of the serial consoles now.