feat!: provide stdout and stderr separately
The guest uses a single serial console to write stdout and stderr. When
using virtrun with go test -exec this is now issue since go test merges
it anyway. But when running without this wrapper being able to handle
the output streams separately may be useful.

This commit splits the output streams by using a separate serial console
for stdout. It uses a base64 encoded pipe on top so it is possible to
transmit any kind of data via stdout. Additionally, the virtrun exit
code is  sent on stderr as well now. The host parses this output stream
only.

The guest's binary stdout is sent on stdout to the user. The qemu
stderr, the guest's kernel log and the guest's binary stderr are merged
and sent on stderr to the user.

BREAKING: The guest's stderr is not sent on stdout anymore.